Test Get SuperUser
