+++ server/src/tests/UserTest/userTest.test.js
@@ -1,0 +1,1 @@
+//Acá iría el test de user
